% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focal_brick.R
\name{focal_brick}
\alias{focal_brick}
\title{Focal Stat brick}
\usage{
focal_brick(brick, window, fun = mean)
}
\arguments{
\item{brick}{a raster brick object}

\item{window}{a matrix defining the windo across which to calculate the statistic}

\item{fun}{the function to use when calculating the statistic (mean by default)}
}
\value{
a raster brick with the same extents as \code{brick}
}
\description{
Calculate focal stats for each layer in a raster brick
}
