% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cover.R
\name{get_cover}
\alias{get_cover}
\title{Get Cover}
\usage{
get_cover(x, raster)
}
\arguments{
\item{x}{an sf object which must be a polygon/multipolygon}

\item{raster}{the model raster to define the output extents and resolution}
}
\value{
a raster with a value of 1 for every cell covered by x
}
\description{
Get the cover of an sf object based upon an input raster
}
