% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_sf_radius.R
\name{max_sf_radius}
\alias{max_sf_radius}
\title{Calculate Max Radius}
\usage{
max_sf_radius(x, numeric = FALSE)
}
\arguments{
\item{x}{an sf object}

\item{numeric}{a logical value defining whether the output should be converted to a numeric value}
}
\value{
either a numeric or units class of the maximum distance from the center of x to it's furthest extent
}
\description{
Calculate the maiximum distance from the center of an sf object to its furthest extent
}
