% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outerBuffer.R
\name{outerBuffer}
\alias{outerBuffer}
\title{Convex hull with buffer}
\usage{
outerBuffer(x, d = 0)
}
\arguments{
\item{x}{\code{sf} polygon object}

\item{d}{optional buffer distance (m)}
}
\value{
\code{sf} polygon object
}
\description{
Draw a convex hull around polygon geometries, creating a single polygon.
}
