% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimeter.R
\name{perimeter}
\alias{perimeter}
\title{Perimeter of multiple polygons}
\usage{
perimeter(x)
}
\arguments{
\item{x}{\code{sf} multipolygons objects}
}
\value{
\code{sf} polygon object
}
\description{
Perimeter of multiple polygons
}
