% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_points.R
\name{random_points}
\alias{random_points}
\title{Generate Random Points}
\usage{
random_points(bbox, n_points, seed = NULL)
}
\arguments{
\item{bbox}{a bounding box object generated by st_bbox}

\item{n_points}{the number of points to generate}

\item{seed}{optionally use set.seed to make the point reproducible}
}
\value{
a sf object of points with the EPSG 27700
}
\description{
Generate n random points within a bounding box
}
