% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_raster.R
\name{sf_to_raster}
\alias{sf_to_raster}
\title{SF to Raster}
\usage{
sf_to_raster(sf, resolution)
}
\arguments{
\item{sf}{and sf object}

\item{resolution}{a numeric defining the resolution of the output raster}
}
\value{
a raster object with the same extents as \code{sf} and the resolution set by \code{resolution}
}
\description{
Generate an empty raster based upon the extents and projection of an sf object
}
