% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialutils-package.R
\docType{package}
\name{spatialutils-package}
\alias{spatialutils}
\alias{spatialutils-package}
\title{spatialutils: Utilities for Working with Spatial Data}
\description{
Miscellaneous utilities to facilitate working with spatial data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FOR-CAST/spatialutils}
  \item Report bugs at \url{https://github.com/FOR-CAST/spatialutils/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

}
\keyword{internal}
