% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_bbox.R
\name{split_bbox}
\alias{split_bbox}
\title{Split Bounding Box}
\usage{
split_bbox(bbox, n_x = 2, n_y = 2)
}
\arguments{
\item{bbox}{a bounding_box created by sf::st_bbox()}

\item{n_x}{the number of boxes on the x axes}

\item{n_y}{the number of boxes on the y axis}
}
\value{
a list of bounding boxs generated by sf::st_bbox()
}
\description{
Split a bounding box into an equal number of squares
}
