% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_work.R
\name{split_work}
\alias{split_work}
\title{Define parallel work splits}
\usage{
split_work(length, cores)
}
\arguments{
\item{length}{the length of an object to split (nrows, list length, etc)}

\item{cores}{the number of cores available on your machine. Typically \code{parallel::detectCores() - 1}.}
}
\value{
a list of length equal to the value of \code{cores}
}
\description{
Define an index of work splits for parallel processing
}
\details{
This function accepts an integar defining the length of you object to be split and will return a list
of length equal to n_cores. This can be used to split up an  object for use in parallel processing such
as mclapply
}
