% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_erase.R
\name{st_erase}
\alias{st_erase}
\title{Erase features}
\usage{
st_erase(x, y)
}
\arguments{
\item{x, y}{sf polygon objects}
}
\value{
sf polygon object
}
\description{
Erase features \code{y} from \code{x}.
Adapted from \url{https://github.com/r-spatial/sf/issues/346}.
}
