% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_grid_stat_parallel.R
\name{st_grid_stat_parallel}
\alias{st_grid_stat_parallel}
\title{Calculate grid stats}
\usage{
st_grid_stat_parallel(x, grid, stat = "length")
}
\arguments{
\item{x}{an sf object}

\item{grid}{an sf object produced by st_make_grid}

\item{stat}{a string specifying which statistic to calculate}
}
\value{
a raster object with the same extent and resolution as \code{grid}
}
\description{
Use paralell processing to calculate grid stats
}
\details{
See \code{get_cover}, which can be used to acheive similar results for some problems at a much better speed
}
