% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_bbox.R
\name{tile_bbox}
\alias{tile_bbox}
\title{Split Bounding Box}
\usage{
tile_bbox(bbox, resolution, ratio = 1)
}
\arguments{
\item{bbox}{a bounding_box created by sf::st_bbox()}

\item{resolution}{the resolution (in the units of the crs) of the output bbox.}

\item{ratio}{the ratio of the x to y dimension of the output bbox}
}
\value{
a list of bounding boxs generated by sf::st_bbox()
}
\description{
Split a bounding box into an equal number of squares
}
