% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_intersection.R
\name{xy_intersection}
\alias{xy_intersection}
\title{Get values from intersecting objects}
\usage{
xy_intersection(x, y, x_id, y_id)
}
\arguments{
\item{x}{an sf object}

\item{y}{an sf object}

\item{x_id}{a column in x that uniquely identfies features}

\item{y_id}{a column in y that uniquely identfies features}
}
\value{
a tibble with nrows equal to nrows(x) containing the intersecting values
}
\description{
Extract values from intersecting sf objects
}
