% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_to_points.R
\name{xy_to_points}
\alias{xy_to_points}
\title{Covert xy to points}
\usage{
xy_to_points(x = 456203, y = 503210, crs = 27700)
}
\arguments{
\item{x}{the x coordinate of the point}

\item{y}{the y coordinate of the point}

\item{crs}{the crs of the point, default is epsg 27700 (BNG)}
}
\value{
an sf object of points
}
\description{
Convert XY coordinates to a points object based upon xy coordinates
}
