% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{calc_centroid}
\alias{calc_centroid}
\title{Get the coordinates of the centroid of a polygon}
\usage{
calc_centroid(poly)
}
\arguments{
\item{poly}{An \code{sf} or \code{sfc} polygon object.}
}
\value{
A numeric vector of centroid coordinates (\code{x}, \code{y}).
}
\description{
Computes the centroid of a polygon and returns its coordinates as a numeric
vector. The attribute-geometry relationship is set to \code{"constant"} to
suppress \code{\link[sf:st_centroid]{sf::st_centroid()}} warnings.
}
\examples{
poly <- sf::st_sf(
  id = 1,
  geometry = sf::st_sfc(
    sf::st_polygon(list(rbind(c(0, 0), c(2, 0), c(2, 2), c(0, 2), c(0, 0)))),
    crs = 4326
  )
)
calc_centroid(poly)
}
\seealso{
Other spatial helpers:
\code{\link[=create_bbox]{create_bbox()}},
\code{\link[=extract_coords]{extract_coords()}},
\code{\link[=prep_agg_lonlat_raster]{prep_agg_lonlat_raster()}},
\code{\link[=save_gpkg]{save_gpkg()}},
\code{\link[=spatial_join_intersects]{spatial_join_intersects()}}
}
\concept{spatial helpers}
