% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{create_bbox}
\alias{create_bbox}
\title{Create a bounding box of a study area}
\usage{
create_bbox(studyArea)
}
\arguments{
\item{studyArea}{An \code{sf} or \code{sfc} object (or anything accepted by
\code{\link[sf:st_bbox]{sf::st_bbox()}}).}
}
\value{
An object of class \code{bbox} (see \code{\link[sf:st_bbox]{sf::st_bbox()}}).
}
\description{
Convenience wrapper around \code{\link[sf:st_bbox]{sf::st_bbox()}} used to derive a bounding box
from a study area, e.g. to reduce processing time of large vector datasets.
}
\examples{
poly <- sf::st_sfc(
  sf::st_polygon(list(rbind(c(0, 0), c(1, 0), c(1, 1), c(0, 1), c(0, 0)))),
  crs = 4326
)
create_bbox(poly)
}
\seealso{
Other spatial helpers:
\code{\link[=calc_centroid]{calc_centroid()}},
\code{\link[=extract_coords]{extract_coords()}},
\code{\link[=prep_agg_lonlat_raster]{prep_agg_lonlat_raster()}},
\code{\link[=save_gpkg]{save_gpkg()}},
\code{\link[=spatial_join_intersects]{spatial_join_intersects()}}
}
\concept{spatial helpers}
