% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_clean.R
\name{intersect_clean}
\alias{intersect_clean}
\title{Intersect and cleanup crumbs}
\usage{
intersect_clean(x, y, xcol, areaThresh = 0.05)
}
\arguments{
\item{x, y}{\code{sf} polygons object}

\item{xcol}{character, name of the attribute column in \code{x} to use to identify features}

\item{areaThresh}{numeric, \emph{proportion} of minimum polygon area to use as threshold crumb size.}
}
\value{
\code{sf} polygons object
}
\description{
Useful in cases where two sets of polygons do not align perfectly spatially.
Thus, intersecting produces slivers that are larger than can typically
be easily dealt with (e.g., \url{https://github.com/r-spatial/sf/issues/547}).
}
\details{
In these cases, deal with each of the set of features individually:
\enumerate{
\item Identify the feature sets of interest in \code{x} using the unique values of \code{x[[xcol]]}
(i.e., sets of polygons that share an identifier);
\item For each of the feature sets, calculate the area of the smallest feature in that set;
\item For each of the feature sets, perform intersection with \code{y} and ensure validity;
\item Post-intersection, remove any fragment polygons smaller than \code{areaThresh * min(area)}.
}
}
