% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{spatial_join_intersects}
\alias{spatial_join_intersects}
\title{Spatial join based on intersection}
\usage{
spatial_join_intersects(x, y)
}
\arguments{
\item{x}{An \code{sf} (multi)polygon object to subset.}

\item{y}{An \code{sf} object to intersect against.}
}
\value{
An \code{sf} object: the subset of (cast) features of \code{x} that intersect
\code{y}.
}
\description{
Casts \code{x} to \code{POLYGON} geometries and subsets it to those features that
intersect \code{y}, using \code{\link[sf:st_intersects]{sf::st_intersects()}} as the join predicate. The
attribute-geometry relationship of \code{x} is set to \code{"constant"} to suppress
\code{\link[sf:st_cast]{sf::st_cast()}} warnings.
}
\examples{
grid <- sf::st_sf(
  id = 1:2,
  geometry = sf::st_sfc(
    sf::st_polygon(list(rbind(c(0, 0), c(1, 0), c(1, 1), c(0, 1), c(0, 0)))),
    sf::st_polygon(list(rbind(c(2, 2), c(3, 2), c(3, 3), c(2, 3), c(2, 2)))),
    crs = 4326
  )
)
pt <- sf::st_sf(
  geometry = sf::st_sfc(sf::st_point(c(0.5, 0.5)), crs = 4326)
)
spatial_join_intersects(grid, pt)
}
\seealso{
Other spatial helpers:
\code{\link[=calc_centroid]{calc_centroid()}},
\code{\link[=create_bbox]{create_bbox()}},
\code{\link[=extract_coords]{extract_coords()}},
\code{\link[=prep_agg_lonlat_raster]{prep_agg_lonlat_raster()}},
\code{\link[=repair_geoms]{repair_geoms()}},
\code{\link[=save_gpkg]{save_gpkg()}},
\code{\link[=sym_difference]{sym_difference()}}
}
\concept{spatial helpers}
